requirements.txt
boto3==1.28.0
pandas==2.0.0
pyarrow==12.0.0
fastparquet==2023.4.0
requests==2.31.0
apache-airflow-providers-amazon